feat: add re-usable status screen (#7019)
