+++ src/pages/RFOX/components/Stake/StakeStatus.tsx
@@ -1,15 +1,9 @@
 import { CheckCircleIcon, WarningIcon } from '@chakra-ui/icons'
-import { Button, CardBody, CardFooter, Center, Heading, Link, Stack } from '@chakra-ui/react'
 import { fromAccountId } from '@shapeshiftoss/caip'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
-import { AnimatePresence } from 'framer-motion'
 import React, { useCallback, useMemo } from 'react'
-import { useTranslate } from 'react-polyglot'
 import { useHistory } from 'react-router'
 import { CircularProgress } from 'components/CircularProgress/CircularProgress'
-import { SlideTransition } from 'components/SlideTransition'
-import { SlideTransitionY } from 'components/SlideTransitionY'
-import { Text } from 'components/Text'
 import type { TextPropTypes } from 'components/Text/Text'
 import { getTxLink } from 'lib/getTxLink'
 import { fromBaseUnit } from 'lib/math'
@@ -17,6 +11,7 @@ import { selectAssetById, selectTxById } from 'state/slices/selectors'
 import { serializeTxIndex } from 'state/slices/txHistorySlice/utils'
 import { useAppSelector } from 'state/store'
 
+import { SharedStatus } from '../Shared/SharedStatus'
 import type { RfoxStakingQuote } from './types'
 import { StakeRoutePaths, type StakeRouteProps } from './types'
 
@@ -36,7 +31,6 @@ export const StakeStatus: React.FC<StakeRouteProps & StakeStatusProps> = ({
   txId,
 }) => {
   const history = useHistory()
-  const translate = useTranslate()
 
   const handleGoBack = useCallback(() => {
     history.push(StakeRoutePaths.Input)
@@ -102,31 +96,5 @@ export const StakeStatus: React.FC<StakeRouteProps & StakeStatusProps> = ({
     [stakingAsset?.explorerTxLink, txId],
   )
 
-  return (
-    <SlideTransition>
-      {bodyContent && (
-        <AnimatePresence mode='wait'>
-          <SlideTransitionY key={bodyContent.key}>
-            <CardBody py={12}>
-              <Center flexDir='column' gap={4}>
-                {bodyContent.element}
-                <Stack spacing={0} alignItems='center'>
-                  <Heading as='h4'>{translate(bodyContent.title)}</Heading>
-                  <Text translation={bodyContent.body} />
-                </Stack>
-              </Center>
-            </CardBody>
-          </SlideTransitionY>
-        </AnimatePresence>
-      )}
-      <CardFooter flexDir='column' gap={2}>
-        <Button as={Link} href={txLink} size='lg' variant='ghost' isExternal>
-          {translate('trade.viewTransaction')}
-        </Button>
-        <Button size='lg' colorScheme='blue' onClick={handleGoBack}>
-          {translate('common.goBack')}
-        </Button>
-      </CardFooter>
-    </SlideTransition>
-  )
+  return <SharedStatus onBack={handleGoBack} txLink={txLink} body={bodyContent} />
 }
